Import of WSNYDER/SystemC-Vregs-1.250 from CPAN.
gitpan-cpan-distribution: SystemC-Vregs
gitpan-cpan-version:      1.250
gitpan-cpan-path:         WSNYDER/SystemC-Vregs-1.250.tar.gz
gitpan-cpan-author:       WSNYDER
gitpan-cpan-maturity:     released

+++ include/VregsRegInfo.cpp
@@ -1,7 +1,7 @@
-// $Revision: #15 $$Date: 2004/01/27 $$Author: wsnyder $ -*- C++ -*-
+// $Revision: 1.18 $$Date: 2005/01/12 21:35:09 $$Author: wsnyder $ -*- C++ -*-
 //======================================================================
 //
-// Copyright 2001-2004 by Wilson Snyder.  This program is free software;
+// Copyright 2001-2005 by Wilson Snyder.  This program is free software;
 // you can redistribute it and/or modify it under the terms of either the GNU
 // General Public License or the Perl Artistic License.
 // 
@@ -138,7 +138,7 @@ const char* VregsRegInfo::addr_name (
     }
     bufp += strsize; length -= strsize;
     if (addr != rep->address()) {
-	snprintf (bufp, length, "+%llx", (uint64_t)(addr - rep->address()));
+	snprintf (bufp, length, "+%llx", (unsigned long long)(addr - rep->address()));
     }
     return buffer;
 }
